sdk: fix listdir regression and keep jfs_listXattr2 ABI unchanged
`listdir()` switched its size out-parameter to `c_int64()` but kept
`string_at(buf, size)`. `string_at` declares its second argument as
`c_int`, so ctypes rejects a `c_int64` instance and every `listdir()`
call raised `ArgumentError`, which also broke the fsspec `ls()` path.
Pass `size.value` instead.

`jfs_listXattr2` narrowed its size out-parameter from `*int` to `*C.int`,
which changes the generated header from `GoInt* size` (8 bytes) to
`int* size` (4 bytes). A C caller built against the previous header keeps
reading 8 bytes, so the upper half retains stack garbage after only 4
bytes are written. Use `*int64` instead: `GoInt` and `GoInt64` are both
`long long` in the generated header, so the ABI is unchanged while the
width becomes explicit and consistent with `jfs_listdir2`. The Python
`listxattr()` caller is widened to match.

Add `ListdirTests` covering empty directories, long and non-ASCII names,
`detail=True` stat fields, and many entries; these fail before the fix.

+++ sdk/java/libjfs/main.go
@@ -1219,7 +1219,7 @@ func jfs_listXattr(pid int64, h int64, path *C.char, buf uintptr, bufsize int32)
 }
 
 //export jfs_listXattr2
-func jfs_listXattr2(pid int64, h int64, path *C.char, value **C.char, size *C.int) int32 {
+func jfs_listXattr2(pid int64, h int64, path *C.char, value **C.char, size *int64) int32 {
 	*value = nil
 	*size = 0
 	w := F(h)
@@ -1229,7 +1229,7 @@ func jfs_listXattr2(pid int64, h int64, path *C.char, value **C.char, size *C.in
 	t, err := w.ListXattr(w.withPid(pid), C.GoString(path))
 	if err == 0 {
 		*value = C.CString(string(t))
-		*size = C.int(len(t))
+		*size = int64(len(t))
 	}
 	return errno(err)
 }
